requirements.txt
slackclient
python-dotenv~=0.19.2
flask~=2.0.2
slackeventsapi~=3.0.0
slack
boto3